requirements.txt
streamlit
google-api-python-client
openai
pillow
moviepy
requests
python-dotenv
gtts

